Move test into tsserver/projects.ts
